Solution to Check if a Number is Square of Another

+++ Check if a Number is Square of Another@@6788dd63490637cfc44e1ba4/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
-#include<math.h>
+#include <math.h>
 int main() {
     int a,b,result;
     scanf{"%d %d",&a,&b};
-    result=pow(b,2);
+    result=(int)pow(b,2);
     if(a==result){
         printf("Yes");
     }
